lux: custom power HAL
The main purpose of this HAL is to provide configurable power
profiles. It operates primarily by configuring the behaviour
of the interactive governor to suit the power profile. However,
this HAL is designed to not conflict with alternative governors.
For CPU touch boosting, it uses a booster in the kernel instead
of using the performance hint mechanism in userspace. This
simplifies the HAL and provides compatibility with alternate
governors.

Change-Id: I5096cdb4bcda2b0c9adb2024198e3a57a9b72063